Rename 알고리즘 스터디 - 숫자 카드 게임(그리디).cpp to 숫자 카드 게임(그리디).cpp
